+++ core.h
@@ -378,7 +378,9 @@ TermCriteria TermCriteria_New(int typ, int maxCount, double epsilon);
 int64_t GetCVTickCount();
 double GetTickFrequency();
 
+AsyncArray AsyncArray_New();
 void AsyncArray_GetAsync(AsyncArray async_out,Mat out);
+void AsyncArray_Close(AsyncArray a);
 
 #ifdef __cplusplus
 }
